Integra saldo para atualizar conforme transações
