+++ packages/byte-bank/src/app/api/account/balance/route.ts
@@ -1,8 +1,8 @@
 import request from "$/app/api/config";
 import mock from "$/app/api/mock.json";
 
-export async function GET(): Promise<Response>  {
-  const data = { balance: mock.account.balance };
+export async function GET(): Promise<Response> {
+  const data = { balance: mock.account.transactions.reduce(((acc, cur) => acc += cur.value), 0) };
 
   return request({
     data,
